+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
+## 0.1.1 - 2020-01-29
+### Added
+* `update`, `field-delete`, `increment`, `server-timestamp`, `array-union`, `array-remove`, `get-id`
+
+### Modified
+- `->atom` now returns an atom instead of a promise.
+
 ## 0.1.0 - 2020-01-28
 ### Added
 - First release